Moved model utils into correct place
