+++ tests/meta.test.ts
@@ -15,10 +15,10 @@ import {
   RECORD_TIMESTAMP_REGEX,
   queryEphemeralDatabase,
 } from '@/fixtures/utils';
+import { getSystemFields } from '@/src/model';
 import { slugToName } from '@/src/model/defaults';
 import type { MultipleRecordResult } from '@/src/types/result';
 import { QUERY_SYMBOLS, RoninError } from '@/src/utils/helpers';
-import { getSystemFields } from '@/src/utils/model';
 
 test('create new model', () => {
   const fields: Model['fields'] = [
